requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
flatbuffers==1.12
gast==0.4.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
h5py==3.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pykospacing @ git+https://github.com/haven-jeon/PyKoSpacing.git@d1e2f93759b1bcc74b6aaa345a62c08701e5546e
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
xlwt==1.3.0
youtube-transcript-api==0.6.1
zipp==3.16.2
